app[REFACT] moved signInFetch in VM
